refactor(admin): move settings stores to runes
